backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
pydantic==1.10.13
uvicorn==0.24.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1
sqlalchemy==1.4.41
python-dotenv==1.0.0
httpx==0.25.2
bcrypt==4.0.1